Moved some base cache classes (#79)
