+++ src/services/cache/DheJsApiCache.ts
@@ -1,7 +1,7 @@
 import { initDheApi } from '@deephaven/require-jsapi';
 import type { EnterpriseDhType as DheType } from '@deephaven-enterprise/jsapi-types';
 import { getTempDir, urlToDirectoryName } from '../../util';
-import { CacheByUrlService } from '../CacheByUrlService';
+import { CacheByUrlService } from './CacheByUrlService';
 
 /**
  * Cache DHE jsapi instances by URL.
