+++ src/services/cache/index.ts
@@ -1,3 +1,4 @@
+export * from './CacheByUrlService';
 export * from './DheClientCache';
 export * from './DheJsApiCache';
 export * from './DheServiceCache';
